requirements.txt
fastapi 
uvicorn  
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
python-dotenv
email-validator
cryptography
websockets